refactor: decompose responsibilities into focused modules
Extract test environment detection from LogFilter into TestEnvironmentDetector,
extract console output routing into LogOutput, and split constants.ts into
domain-specific modules. Logger pipeline is now an explicit 6-stage flow:
detect → filter(level) → filter(key) → build → format → output.

Public API unchanged. All tests pass.
